requirements.txt
# LendenClub Voice Assistant - Phase 2 Requirements
# Updated requirements including RAG engine and enhanced scraping

# Core ML/AI Libraries
transformers>=4.30.0          # For BART-Large-MNLI and other models
torch>=2.0.0                 # PyTorch backend for transformers
sentence-transformers>=2.2.0  # For document embeddings in RAG
faiss-cpu>=1.7.4             # Vector database for document search
scikit-learn>=1.3.0          # Traditional ML algorithms and metrics

# Web Scraping Enhancement  
requests>=2.31.0             # HTTP requests
beautifulsoup4>=4.12.0       # HTML parsing
fake-useragent>=1.4.0        # User agent rotation for anti-detection
lxml>=4.9.0                  # Fast XML parser for BeautifulSoup

# Data Processing
pandas>=2.0.0                # Data manipulation and analysis
numpy>=1.24.0                # Numerical computations
python-dateutil>=2.8.0       # Date/time utilities

# Visualization and Reporting
matplotlib>=3.7.0            # Plotting and visualization
seaborn>=0.12.0              # Statistical visualizations
plotly>=5.15.0               # Interactive plots (optional)

# Logging and Configuration
python-dotenv>=1.0.0         # Environment variable management
pyyaml>=6.0                  # YAML configuration files

# Testing Framework
pytest>=7.4.0               # Testing framework
pytest-cov>=4.1.0           # Coverage reporting

# Development Tools
black>=23.0.0                # Code formatting
flake8>=6.0.0                # Code linting
mypy>=1.5.0                  # Type checking

# Optional GPU Support (uncomment if you have CUDA GPU)
# torch>=2.0.0+cu118          # PyTorch with CUDA support
# faiss-gpu>=1.7.4            # GPU-accelerated FAISS

# Optional Advanced Features
# streamlit>=1.25.0           # For web interface (Phase 3)
# fastapi>=0.100.0            # API framework (Phase 4)
# uvicorn>=0.23.0             # ASGI server (Phase 4)